requirements.txt
tensorflow==2.15.0
pydicom==2.4.4
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
opencv-python==4.9.0.80
pillow==10.2.0
scikit-image==0.19.1
gdown==5.1.0